backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
numpy==1.26.4
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.4.0
PyMuPDF==1.24.10
xlsxwriter==3.2.0
reportlab==4.2.2
python-multipart==0.0.9
pydantic==2.7.3
jinja2==3.1.4
supabase==2.6.0
